utils independent of application context
