+++ package.json
@@ -28,7 +28,6 @@
     "graphql": "^14.6.0",
     "graphql-tag": "^2.10.3",
     "immer": "^6.0.2",
-    "localforage": "^1.7.3",
     "react": "^16.13.0",
     "react-feather": "^2.0.3",
     "redux": "^4.0.5",
@@ -41,7 +40,6 @@
   "devDependencies": {
     "@types/apollo-upload-client": "^8.1.3",
     "@types/graphql": "^14.5.0",
-    "@types/node": "^13.9.1",
     "@types/react": "^16.9.26",
     "@types/redux-actions": "^2.6.1",
     "@types/reselect": "^2.2.0",
