+++ src/imageValidator.ts
