+++ src/redux/state.ts
@@ -1,6 +1,7 @@
 import { of } from 'rxjs';
 import { handleActions } from 'redux-actions';
 import { produce } from 'immer';
+import gql from 'graphql-tag';
 
 import ActionState from './actions';
 
@@ -19,8 +20,8 @@ export const initialMetaState = {
   delete: defaultState
 };
 
-export const selectEntities = state => state.entities;
-export const selectEntitiesMeta = state => state.entitiesMeta;
+export const entities = state => state.entities;
+export const entitiesMeta = state => state.entitiesMeta;
 
 export function responder(response, context = '') {
   const NO_INTERNET_MESSAGE = 'No internet, please check your network connection and try again';
@@ -35,7 +36,7 @@ export function success$(action, response) {
   return of(action(response.data).success);
 }
 
-export function createMetaReducer(action: ActionState) {
+export function metaReducer(action: ActionState) {
   return (
     handleActions({
       [action.create.loading]: (state) => produce(state, draft => { draft.create.loading = true }),
@@ -84,4 +85,35 @@ export function createMetaReducer(action: ActionState) {
   );
 }
 
+export function gqlResponder(error) {
+  const NO_INTERNET_MESSAGE = 'No internet, please check your network connection and try again';
+  const SERVICE_UNAVAILABLE_MESSAGE = 'Service unavailable, please try again later';
+  const UNEXPECTED_ERROR_MESSAGE = 'Please try again, an unexpected error occured';
 
+  try {
+    const { graphQLErrors, networkError } = error;
+
+    if (networkError) {
+      console.log('Network Error', networkError);
+      return SERVICE_UNAVAILABLE_MESSAGE
+    }
+
+    if (graphQLErrors) {
+      const { code, response } = graphQLErrors[0].extensions;
+      console.log('GraphQL Error', response);
+
+      if (response) {
+        return response.body.data.constructor === Array ? response.body.data[0].message : response.body.data;
+      } else {
+        const error = code == 'ECONNREFUSED' ? SERVICE_UNAVAILABLE_MESSAGE : UNEXPECTED_ERROR_MESSAGE
+        return error;
+      }
+    }
+
+    return UNEXPECTED_ERROR_MESSAGE;
+  } catch (ex) {
+    console.log(ex);
+
+    return ex.message;
+  }
+}
