+++ src/storage/cartStorage.ts
@@ -1,19 +1,20 @@
-import { getItem, setItem } from 'localforage';
+import { windowExists } from '../';
+
 /**
  * default implemetation of locas storage
  */
 export default class LocalStorage {
-  private key;
+  private key: string;
 
-  constructor(key) {
+  constructor(key: string) {
     this.key = key;
   }
 
-  public async get() {
-    return await getItem(this.key);
+  public get() {
+    return windowExists.localStorage.getItem(this.key);
   };
 
-  public async set(data) {
-    return await setItem(this.key, data);
+  public set(data) {
+    return windowExists.localStorage.setItem(this.key, data);
   }
 }
